More cosmetic changes
